# BZ 1207826 - speed up metric pruning