--- a/app/models/metric/purging.rb
+++ b/app/models/metric/purging.rb
@@ -50,26 +50,23 @@ module Metric::Purging
     end
   end
 
-  def self.klass_for_interval(interval)
-    klass, conditions = case interval
-                        when 'realtime' then [Metric,       {}]
-                        else             [MetricRollup, {:capture_interval_name => interval}]
-                        end
+  def self.scope_for_interval(interval)
+    interval == 'realtime' ? Metric : MetricRollup.where(:capture_interval_name => interval)
   end
 
   def self.purge_count(older_than, interval)
-    klass, conditions = klass_for_interval(interval)
+    scope = scope_for_interval(interval)
 
-    oldest = klass.select(:timestamp).where(conditions).order(:timestamp).first
+    oldest = scope.select(:timestamp).order(:timestamp).first
     oldest = oldest.nil? ? older_than : oldest.timestamp
 
-    klass.where(conditions).where(:timestamp => oldest..older_than).count
+    scope.where(:timestamp => oldest..older_than).count
   end
 
   def self.purge(older_than, interval, window = nil, limit = nil)
     _log.info("Purging #{limit.nil? ? "all" : limit} #{interval} metrics older than [#{older_than}]...")
 
-    klass, conditions = klass_for_interval(interval)
+    scope = scope_for_interval(interval)
 
     total = 0
     total_tag_values = 0
@@ -78,13 +75,13 @@ module Metric::Purging
 
       oldest = nil
       Benchmark.realtime_block(:query_oldest) do
-        oldest = klass.select(:timestamp).where(conditions).order(:timestamp).first
+        oldest = scope.select(:timestamp).order(:timestamp).first
         oldest = oldest.nil? ? older_than : oldest.timestamp
       end
 
       loop do
         batch, = Benchmark.realtime_block(:query_batch) do
-          klass.select(:id).where(conditions).where(:timestamp => (oldest..older_than)).limit(window)
+          scope.select(:id).where(:timestamp => (oldest..older_than)).limit(window)
         end
         break if batch.empty?
 
@@ -93,7 +90,7 @@ module Metric::Purging
 
         _log.info("Purging #{ids.length} #{interval} metrics.")
         count, = Benchmark.realtime_block(:purge_metrics) do
-          klass.delete_all(:id => ids)
+          scope.unscoped.delete_all(:id => ids)
         end
         total += count
 
@@ -104,7 +101,7 @@ module Metric::Purging
           _log.info("Purging associated tag values.")
           ids.each_slice(50) do |vp_ids|
             tv_count, = Benchmark.realtime_block(:purge_vim_performance_tag_values) do
-              VimPerformanceTagValue.delete_all(:metric_id => vp_ids, :metric_type => klass.name)
+              VimPerformanceTagValue.delete_all(:metric_id => vp_ids, :metric_type => scope.name)
             end
             count_tag_values += tv_count
             total_tag_values += tv_count