This is related to https://bugzilla.redhat.com/show_bug.cgi?id=1207826 and [1226429](https://bugzilla.redhat.com/show_bug.cgi?id=1226429)

This simplifies the Cap&U metric purging code. Goal is to reduce the cost of pruning, so we don't have to introduce different batching mechanisms.

From here, we can possibly move more of the id fetching into a temp table or something.

/cc @dmetzger57 once our log gets lower, I want to improve so we don't bring back so many records from the db
/cc @akrzos fyi - and thanks

area|before|after
---|---|---
metrics deleted|11,222|11,222
tags deleted|0|0
code|231ms|115ms
sql|288ms|215ms
queries|40|27
memory|18,102kb|6,508kb
objects|206,528|132,625

---

Timings from the log

name|before|after
---|---|---
query_oldest|0.0630779|
query_batch|0.0027899|0.095520
purge_metrics|0.2500097|0.241861
total_time|0.8555488|0.558689