@Fryguy 2015-11-19
Also please put the BZ in the commit message themselves.  The bot can't do anything with the commits with respect to commenting on a BZ if it's not in the commit.