@kbrock 2015-11-20
@Fryguy Is there a deeper reason behind `:timestamp => oldest..older_than` vs `[timestamp <= ?', older_than]`? I assumed it was just due to how code changes over time, but wanted to make sure it wasn't there to fix some metrics sub table issue.

Think the biggest savings here was inadvertently removing a number of `count` queries.