@akrzos 2015-11-20
@kbrock I assume you have more than one before/after sample right?  If not I'm sure I can cook up a db with a large scale of expired metrics.