@kbrock 2015-11-20
@akrzos I ran the pruning once, ran garbage collection and then ran the pruning 3 times (rolling back the transaction each time)

I only used 1 dataset thought. So it is very possible that this is just for that single data set.