@kbrock 2015-11-20
@Fryguy fixed that class reference.
validated got the same timing and number of rows changed