@Fryguy 2015-11-20
The reasoning was that I wanted to prune the oldest records first to prevent gaps in the data.  Because databases are not ordered, deleting with a limit leaves intermittent gaps.  I was concerned this might affect reports that ask for hourly data beyond the pruning window, which they all do because they don't qualify with timestamp.  Those graphs would have anomalous entries at the older end of the graph.

However, given that it completely kills the performance of the purge to use the timestamp sorting, I think it's ok to have it go by id as you have done in this PR.  This change will allow the purger to actually complete and keep up more frequently.  If the purger can keep up, then the problem with the graphs won't exist.