@Fryguy 2015-11-20
@kbrock Just to verify, are there specs around this?  If not, we have specs for the myriad other purgers, and it would not be hard to copy that pattern and create specs for this purger.