@Fryguy 2015-11-20
> Is there a deeper reason behind :timestamp => oldest..older_than vs [timestamp <= ?', older_than]? 

I just read this comment a little deeper and realized I didn't answer it.  To be honest, I don't remember.  It may have be related to the [MySQL issue](http://bugs.mysql.com/bug.php?id=54588) I found on indexed timestamp columns, but I'm not really sure.

Here's the commit that introduced it:

```patch
commit 51520cd91b182f01bf44a5b5980cb6b37078419b
Author: Jason Frey <jfrey@redhat.com>
Date:   Mon Mar 28 19:29:29 2011 +0000

    Changed VimPerformance purger to delete rows in batches based on number of rows, defaulting to 1000, and configurable in vmdb.yml
    BugzID:11868
    
    git-svn-id: http://miq-ubuntusub.manageiq.com/svn/svnrepos/Manageiq/trunk@27598 3c68ef56-dcc3-11dc-9475-a42b84ecc76f

diff --git a/vmdb/app/models/vim_performance.rb b/vmdb/app/models/vim_performance.rb
index c927805..01487ae 100644
--- a/vmdb/app/models/vim_performance.rb
+++ b/vmdb/app/models/vim_performance.rb
@@ -253,27 +253,27 @@ class VimPerformance < ActiveRecord::Base
     )
   end
 
-  def self.purge(older_than, interval, window = nil)
-    oldest = self.find(:first, :conditions => {:capture_interval_name => interval}, :order=> "timestamp ASC", :select => :timestamp)
+  def self.purge(older_than, interval, window = nil, limit = nil)
+    log_header = "MIQ(VimPerformance.purge)"
+    $log.info("#{log_header} Purging #{limit.nil? ? "all" : limit} #{interval} performances older than [#{older_than}]...")
+
+    window ||= (VMDB::Config.new("vmdb").config.fetch_path(:performance, :history, :purge_window_size) || 1000)
+
+    oldest = self.first(:select => :timestamp, :conditions => {:capture_interval_name => interval}, :order => :timestamp)
     oldest = oldest.nil? ? older_than : oldest.timestamp
 
-    $log.info("MIQ(VimPerformance-purge) Pre-Purging Summary: all #{interval} performances older than [#{older_than}], oldest timestamp: [#{oldest}]")
-
-    window ||= (VMDB::Config.new("vmdb").config.fetch_path([:performance, :history, :purge_window]) || 1.month).to_i_with_method
-
-    ts = nil
     total = 0
-    loop do
-      ts = ts.nil? ? (oldest + window) : (ts + window)
-      ts = older_than if (ts > older_than)
+    until (batch = self.all(:select => :id, :conditions => {:capture_interval_name => interval, :timestamp => oldest..older_than}, :limit => window)).empty?
+      ids = batch.collect { |b| b.id }
+      ids = ids[0, limit - total] if limit && total + ids.length > limit
 
-      $log.info("MIQ(VimPerformance-purge) Purging all #{interval} performances older than [#{ts}]...")
-      deletes = self.delete_all(["timestamp < :ts AND capture_interval_name = :interval", {:ts => ts, :interval => interval}])
-      total += deletes
-      $log.info("MIQ(VimPerformance-purge) Purging all #{interval} performances older than [#{ts}]... Complete. Deleted #{deletes} records")
-      break if (ts == older_than)
+      $log.info("#{log_header} Purging #{ids.length} #{interval} performances.")
+      total += self.delete_all(:id => ids)
+
+      break if limit && total >= limit
     end
-    $log.info("MIQ(VimPerformance-purge) Post-Purging Summary: all #{interval} performances older than [#{older_than}]. Deleted #{total} total records ")
+
+    $log.info("#{log_header} Purging #{limit.nil? ? "all" : limit} #{interval} performances older than [#{older_than}]...Complete - Deleted #{total} records")
   end
 
   #
diff --git a/vmdb/config/vmdb.tmpl.yml b/vmdb/config/vmdb.tmpl.yml
index b96e619..7020222 100644
--- a/vmdb/config/vmdb.tmpl.yml
+++ b/vmdb/config/vmdb.tmpl.yml
@@ -331,7 +331,7 @@ performance:
     :keep_realtime_performances: 4.hours
     :keep_hourly_performances: 6.months
     :keep_daily_performances: 6.months
-    :purge_window: 1.month
+    :purge_window_size: 1000
   concurrent_requests:
     :hourly: 1
     :realtime: 20
```