@kbrock 2015-11-20
1. `timestamp between min(timestamp)..oldest = minimum(timestamp) ; :timestamp => oldest..older_than` is functionally identical to `timestamp <= ?, older_than` - so I don't think this code did anything before.
2. There is no `order` by `timestamp` (or `id). We can add that if you want.
3. I do want to revisit and see what it looks like if we keep the `ids` in a Postgres `#temp` table. If it stays readable, it should give us a big win for cpu, io, and ram in both the postgres and ruby processes.

@Fryguy re transaction log - was this more of a problem with MySQL, or do we also have the problem with Postgres. Did we do work to figure out the correct values for the numbers?