@Fryguy 2015-11-20
> There is no order by timestamp (or `id). We can add that if you want.

Noooo that's the performance killer.  It used to be [here](https://github.com/ManageIQ/manageiq/pull/5536/files) and [here](https://github.com/ManageIQ/manageiq/pull/5536/files#diff-305de81cc1b046532b365cc291edccfdL80)

> I do want to revisit and see what it looks like if we keep the ids in a Postgres #temp table

I'd need to understand what you are trying to gain for this.  Keeping 1000 ids in memory should not make a difference.

> re transaction log

I believe this is still an issue in PG, which is one thing the batches tries to help.  The other thing it helps is failures mid purge.  If you wrap 100000s delete in a single transaction and it fails at record 99999, it has to rollback *all* of that, and then it has to start all over again.  Committing more frequently allows it to checkpoint.