@kbrock 2015-11-21
> I'd need to understand what you are trying to gain for this. Keeping 1000 ids in memory should not make a difference.

transporting 1000 ids over the wire from postgres to ruby. Generate a sql statement with all those numbers, to send back over the wire to postgres to then parse the numbers and put them into a set (similar to a temporary table) to then join to a table and delete.

vs

Leave them as a set/temp table in postgres, bypassing most of the steps, especially the transportation (network) and serialization steps (cpu).

This is for another day.